CRUD with auth MW depending on user/admin ready. Waiting for deploy to prod

+++ controllers/ReviewControllers.js
@@ -1,15 +1,23 @@
 const Review = require('../models/review')
+const User = require('../models/user')
+var sequelize = require('../db/db')
+const jsonwebtoken = require("jsonwebtoken");
 
 const ReviewController = {};
 
 ReviewController.postNewReview = async (req, res) => {
     try {
         let data = req.body
+        const { authorization } = req.headers;
+        const [strategy, jwt] = authorization.split(" ");
+        const payload = jsonwebtoken.verify(jwt, process.env.JWT_SECRET)
+        console.log(payload)
+        let mail = payload.mail
         let resp = await Review.create({
             text: data.text,
             rating: data.rating,
             creation_date: data.creation_date,
-            userMail: data.userMail,
+            userMail: mail,
             serviceIdService: data.serviceIdService
         })
 
@@ -21,16 +29,31 @@ ReviewController.postNewReview = async (req, res) => {
 
 ReviewController.deleteReviewById = async (req, res) => {
     try {
+        let data = req.body
+        const { authorization } = req.headers;
+        const [strategy, jwt] = authorization.split(" ");
+        const payload = jsonwebtoken.verify(jwt, process.env.JWT_SECRET)
+        console.log(payload)
+        let mail = payload.mail
+
         let id = req.params.id
-        let resp = await Review.destroy({
+        let user = await Review.findOne({
             where: { id_review: id }
         })
-
-        if (resp == 1) {
-            res.send("Se ha eliminado el review correctamente")
+        if (user.userMail === mail) {
+            let resp = await Review.destroy({
+                where: { id_review: id }
+            })
+            console.log(resp)
+            if (resp == 1) {
+                res.send("Se ha eliminado el review correctamente")
+            } else {
+                res.status(400).send("No se ha podido eliminar el review")
+            }
         } else {
-            res.status(400).send("No se ha podido eliminar el review")
+            console.log("falloo")
         }
+
 
     } catch (err) {
         res.send(err)
